Session Update
Sessions sollten jetzt funktionieren. Man kann sich einloggen und
wieder ausloggen und bekommt jeweils eine neue Session zugewiesen,
zusätzlich wird beim Zugriff auf die Rechte des Users geachtet, normale
User haben nur auf die Cams Zugriff! Es werden auch nur die Cams des
Users angezeigt! Der Admin kann alles machen.

Momentan teilen sich beide eben dieselbe Oberfläche.

+++ WaiWeb/src/servlets/SettingsServlet.java
@@ -2,12 +2,12 @@ package servlets;
 
 import java.io.IOException;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import Dao.DatabaseControllerImpl;
 
@@ -30,21 +30,36 @@ public class SettingsServlet extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String action = request.getParameter("action");
+		HttpSession session = request.getSession(false);
 		
-		if(action.equals("deleteDatabase")){
-			//Komplette Datenbank zurücksetzen TODO: Funktion zum rücksetzen der Datenbanken ohne Admin zu löschen!
-			DatabaseControllerImpl databaseDelete = new DatabaseControllerImpl();
-			databaseDelete.deleteDatabase();
+        if (session != null){
+        	if ((int) session.getAttribute("rechte") == 1) {
+	        	System.out.println("Session mit User=" + session.getAttribute("username") 
+	        			+ " und Rechte=" + session.getAttribute("rechte") + " bestätigt.");
 			
-			backToAuswahl(request, response);
+				if(action.equals("deleteDatabase")){
+					//Komplette Datenbank zurücksetzen TODO: Funktion zum rücksetzen der Datenbanken ohne Admin zu löschen!
+					DatabaseControllerImpl databaseDelete = new DatabaseControllerImpl();
+					databaseDelete.deleteDatabase();
+					
+					backToAuswahl(request, response);
+					
+				} else if(action.equals("deletePictures")){
+					//Alle gespeicherten Bilder löschen: TODO: Funktion um alle Bilder zu löschen!
+					backToAuswahl(request, response);
+					
+				} else if (action.equals("Back")) {
+					backToAuswahl(request, response);
+				} 
+        	} else {
+        		System.out.println("ERROR! Keine ausreichenden Rechte, Administrator-Rechte erforderlich!");
+        		response.sendRedirect(request.getContextPath() + "/master");
+        	}
 			
-		} else if(action.equals("deletePictures")){
-			//Alle gespeicherten Bilder löschen: TODO: Funktion um alle Bilder zu löschen!
-			backToAuswahl(request, response);
-			
-		} else if (action.equals("Back")) {
-			backToAuswahl(request, response);
-		}
+        } else {
+        	System.out.println("ERROR! Keine aktive Session oder ausrechende Rechte gefunden!");
+        	response.sendRedirect(request.getContextPath() + "/master");
+        }
 	}
 
 	/**
@@ -56,8 +71,7 @@ public class SettingsServlet extends HttpServlet {
 	
 	//Funktion um auf Auswahlmöglichkeiten zurückzukehren:
 	void backToAuswahl(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("//jsp/Auswahlmoeglichkeiten.jsp");
-		dispatcher.forward(request, response);
+        response.sendRedirect(request.getContextPath() + "/auswahl");
 	}
 
 }
